+++ extensions/text-shortcuts/src/util/utils.ts
@@ -4,6 +4,7 @@ import Values = Form.Values;
 export const regexPunctuation = /\p{Z}|\p{P}|\p{S}/gu;
 
 export interface Preference extends Values {
+  closeMainWindow: boolean;
   annotation: string;
   caser: string;
   coder: string;