+++ extensions/text-shortcuts/CHANGELOG.md
@@ -1,10 +1,11 @@
 # Text Shortcuts Changelog
 
-## [Update UI Display] - 2022-05-12
+## [Update UI Display] - 2022-05-18
 
 - Release of Text Shortcuts version 1.2
 - Add display of shortcuts tag
 - Add the color display of custom shortcuts
+- Add Preference: Close Main Window
 - Update shortcut keys for delete actions
 
 ## [Update Delete, Replace, Case Actions] - 2022-05-05